Upgrades to RSpec 3
I've updated the gemspec dependency for RSpec to target ~> 3.0 because we should avoid loose dependencies.
But I did this simply because I was using RSpec 3 and needed to get rid of the following deprecation warning:

```
The :example_group key in an example group's metadata hash is deprecated. Use the example group's hash directly for the computed keys and :parent_example_group to access the parent example group metadata instead.
```